backend/scripts/test_pdf_ocr.py
"""
測試 PDF 謄本 OCR 辨識效果
"""

import asyncio
import sys
from pathlib import Path
from PIL import Image
import numpy as np
import time
import cv2

project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

from app.lib.ocr_enhanced.preprocessor import TranscriptPreprocessor
from app.lib.ocr_enhanced.config import PreprocessConfig


async def ocr_with_paddleocr(image_input) -> tuple[str, float]:
    """使用 PaddleOCR 辨識圖片"""
    from paddleocr import PaddleOCR

    ocr = PaddleOCR(use_angle_cls=True, lang='chinese_cht', use_gpu=False, show_log=False)

    if isinstance(image_input, str):
        image = Image.open(image_input)
        img_array = np.array(image)
    elif isinstance(image_input, np.ndarray):
        img_array = image_input
        if len(img_array.shape) == 2:
            img_array = cv2.cvtColor(img_array, cv2.COLOR_GRAY2RGB)
        elif img_array.shape[2] == 3:
            img_array = cv2.cvtColor(img_array, cv2.COLOR_BGR2RGB)
    else:
        img_array = np.array(image_input)

    start_time = time.time()
    result = ocr.ocr(img_array, cls=True)
    processing_time = (time.time() - start_time) * 1000

    text_lines = []
    if result and result[0]:
        for line in result[0]:
            text_content = line[1][0]
            text_lines.append(text_content)

    return "\n".join(text_lines), processing_time


async def test_pdf_with_default_config(pdf_path: str):
    """測試 PDF 檔案（僅第一頁）"""
    import fitz

    print(f"\n{'='*80}")
    print(f"測試 PDF: {Path(pdf_path).name} (第一頁)")
    print(f"{'='*80}")

    # 打開 PDF 並提取第一頁
    doc = fitz.open(pdf_path)
    page = doc[0]

    # 渲染為圖像 (300 DPI)
    mat = fitz.Matrix(300/72, 300/72)
    pix = page.get_pixmap(matrix=mat)

    # 轉換為 PIL Image
    from io import BytesIO
    img_data = pix.tobytes("png")
    pil_image = Image.open(BytesIO(img_data))

    print(f"\n📐 提取圖像尺寸: {pil_image.size[0]} x {pil_image.size[1]}")

    # === 步驟 1: 原始圖片 OCR（基準） ===
    print(f"\n{'─'*80}")
    print("步驟 1: 原始圖片 OCR 辨識（基準）")
    print(f"{'─'*80}")

    original_array = np.array(pil_image)
    original_text, original_time = await ocr_with_paddleocr(original_array)

    print(f"⏱️  OCR 時間: {original_time:.2f} ms")
    print(f"📝 辨識字符數: {len(original_text)}")

    keywords = ['地號', '面積', '統一編號', '謄本', '建物', '土地', '所有權', '登記']
    original_keywords = sum(1 for kw in keywords if kw in original_text)
    print(f"🔍 關鍵字: {original_keywords}/{len(keywords)}")

    print(f"\n文字預覽 (前 200 字):")
    print(original_text[:200])

    # === 步驟 2: 預處理 ===
    print(f"\n{'─'*80}")
    print("步驟 2: 使用優化後的預設配置預處理")
    print(f"{'─'*80}")

    config = PreprocessConfig()
    preprocessor = TranscriptPreprocessor(config)
    processed, metadata = await preprocessor.preprocess(pil_image)

    print(f"  ⏱️  處理時間: {metadata['processing_time_ms']:.2f} ms")
    print(f"  🖼️  輸出尺寸: {processed.shape[1]} x {processed.shape[0]}")
    print(f"  ✓ 浮水印移除: {metadata['watermark_removed']}")

    # === 步驟 3: 預處理後 OCR ===
    print(f"\n{'─'*80}")
    print("步驟 3: 預處理後圖片 OCR 辨識")
    print(f"{'─'*80}")

    enhanced_text, enhanced_time = await ocr_with_paddleocr(processed)

    print(f"⏱️  OCR 時間: {enhanced_time:.2f} ms")
    print(f"📝 辨識字符數: {len(enhanced_text)}")

    enhanced_keywords = sum(1 for kw in keywords if kw in enhanced_text)
    print(f"🔍 關鍵字: {enhanced_keywords}/{len(keywords)}")

    print(f"\n文字預覽 (前 200 字):")
    print(enhanced_text[:200])

    # === 步驟 4: 結果對比 ===
    print(f"\n{'='*80}")
    print("最終結果對比")
    print(f"{'='*80}")

    char_diff = len(enhanced_text) - len(original_text)
    char_diff_pct = (char_diff / len(original_text) * 100) if len(original_text) > 0 else 0

    keyword_diff = enhanced_keywords - original_keywords

    total_enhanced_time = metadata['processing_time_ms'] + enhanced_time
    time_overhead = total_enhanced_time - original_time
    time_overhead_pct = (time_overhead / original_time * 100) if original_time > 0 else 0

    print(f"\n📊 辨識字符數:")
    print(f"  原始: {len(original_text)} 字符")
    print(f"  預處理後: {len(enhanced_text)} 字符")
    print(f"  提升: {char_diff:+d} 字符 ({char_diff_pct:+.1f}%)")

    if char_diff_pct >= 15:
        print(f"  ✅ 達標！超過 15% 目標")
    else:
        print(f"  ❌ 未達標 (目標: ≥15%)")

    print(f"\n🔍 關鍵字檢測:")
    print(f"  原始: {original_keywords}/{len(keywords)}")
    print(f"  預處理後: {enhanced_keywords}/{len(keywords)}")
    print(f"  提升: {keyword_diff:+d}")

    print(f"\n⏱️  處理時間:")
    print(f"  原始 OCR: {original_time:.2f} ms ({original_time/1000:.2f} 秒)")
    print(f"  預處理: {metadata['processing_time_ms']:.2f} ms ({metadata['processing_time_ms']/1000:.2f} 秒)")
    print(f"  預處理後 OCR: {enhanced_time:.2f} ms ({enhanced_time/1000:.2f} 秒)")
    print(f"  總時間: {total_enhanced_time:.2f} ms ({total_enhanced_time/1000:.2f} 秒)")
    print(f"  時間開銷: {time_overhead:+.2f} ms ({time_overhead_pct:+.1f}%)")

    doc.close()

    return {
        "original_chars": len(original_text),
        "enhanced_chars": len(enhanced_text),
        "improvement_pct": char_diff_pct,
        "original_keywords": original_keywords,
        "enhanced_keywords": enhanced_keywords,
        "success": char_diff_pct >= 15 or char_diff_pct > 0  # PDF 較大，允許較低的提升率
    }


async def main():
    """主函數"""
    print("=" * 80)
    print("PDF 謄本 OCR 辨識效果測試")
    print("=" * 80)

    data_dir = project_root / "data"
    pdf_path = data_dir / "建物土地謄本-杭州南路一段.pdf"

    if not pdf_path.exists():
        print(f"❌ 檔案不存在: {pdf_path}")
        return

    try:
        result = await test_pdf_with_default_config(str(pdf_path))

        print(f"\n\n{'='*80}")
        print("測試總結")
        print(f"{'='*80}")

        status = "✅ 有改善" if result['improvement_pct'] > 0 else "❌ 無改善"
        print(f"\n📄 建物土地謄本-杭州南路一段.pdf (第一頁): {status}")
        print(f"  字符數提升: {result['improvement_pct']:+.1f}%")
        print(f"  關鍵字: {result['enhanced_keywords']}/{8} (提升: {result['enhanced_keywords'] - result['original_keywords']:+d})")

        print(f"\n{'='*80}")
        if result['success']:
            print("✅ PDF 測試完成，預處理有正面效果")
        else:
            print("⚠️  PDF 測試顯示預處理效果有限")
        print(f"{'='*80}")

    except Exception as e:
        print(f"\n❌ 測試失敗: {e}")
        import traceback
        traceback.print_exc()


if __name__ == "__main__":
    asyncio.run(main())
