backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.6.1

# AI Services
openai==1.54.5
anthropic==0.39.0

# AWS Services
boto3==1.35.60
botocore==1.35.60

# OCR (Optional)
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
PyMuPDF==1.24.13  # PDF 處理（import 為 fitz）

# PaddleOCR (Better Chinese OCR)
# 版本鎖定以避免 ARM64 Docker 兼容性問題
paddlepaddle==2.6.2
paddleocr==2.8.1
numpy<2.0.0
opencv-python-headless>=4.5.0  # headless 版本避免 GUI 依賴問題

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
